app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
PyPDF2==3.0.1
reportlab==4.0.4
openai==0.27.8
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.24.0
